refactor: migrate to TypeScript and restructure CLI feat: add commit hash to lesson meta
